+++ linera-explorer/html/app.html
@@ -1,0 +1,92 @@
+<nav class="navbar navbar-expand-sm m-2">
+  <div class="container-fluid">
+    <a class="navbar-brand" type="button" @click="route('')">Linera Explorer</a>
+    <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+      <span class="navbar-toggler-icon"></span>
+    </button>
+    <div class="collapse navbar-collapse" id="navbarSupportedContent">
+      <ul class="navbar-nav me-auto">
+      </ul>
+      <ul class="navbar-nav ms-auto">
+        <li class="nav-item mx-2">
+          <input v-model="node" class="form-control">
+        </li>
+        <li class="nav-item mx-2">
+          <select v-model="chain" @change="route()" class="form-select">
+            <option v-for="c in chains" :value="c" :selected="c==chain">{{ sh(c) }}</option>
+          </select>
+        </li>
+        <li class="nav-itemmx-2">
+          <button @click="route()" class="btn btn-primary">
+            <i class="bi bi-arrow-clockwise"></i>
+          </button>
+        </li>
+      </ul>
+    </div>
+  </div>
+</nav>
+<div class="container">
+  <div v-if="page=='home'">
+    <v-block v-if="blocks.length!=0" class="m-2" :block="blocks[0]" title="Head" :route="route">
+    </v-block>
+    <div>
+      <table class="table">
+        <thead>
+          <th>Height</th>
+          <th>Hash</th>
+          <th>Timestamp</th>
+          <th>Signer</th>
+          <th>Status</th>
+          <th>Round</th>
+          <th>#InMessages</th>
+          <th>#OutMessages</th>
+          <th>#Operations</th>
+          <th>JSON</th>
+        </thead>
+        <tbody>
+          <tr v-for="b in blocks">
+            <td>{{ b.value.executedBlock.block.height }}</td>
+            <td :title="b.hash">
+              <a @click="route('block', [['hash', b.hash]])" class="btn btn-link">{{ sh(b.hash) }}
+            </td>
+            <td>{{ (new Date(b.value.executedBlock.block.timestamp/1000)).toLocaleString() }}</td>
+            <td :title="b.value.executedBlock.block.authenticatedSigner">{{ sh(b.value.executedBlock.block.authenticatedSigner) }}</td>
+            <td>{{ b.value.status }}</td>
+            <td>{{ b.value.round }}</td>
+            <td>{{ b.value.executedBlock.block.incomingMessages.length }}</td>
+            <td>{{ b.value.executedBlock.messages.length }}</td>
+            <td>{{ b.value.executedBlock.block.operations.length }}</td>
+            <td>
+              <button class="btn btn-link btn-sm" data-bs-toggle="modal" :data-bs-target="'#json-block-'+b.hash">
+                <i class="bi bi-braces"></i>
+              </button>
+              <div :id="'json-block-'+b.hash" class="modal fade">
+                <div class="modal-dialog modal-xl">
+                  <div class="modal-content">
+                    <div class="modal-header">
+                      <h5 class="modal-title">Block</h5>
+                      <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
+                    </div>
+                    <div class="modal-body">
+                      <v-json :data="b" :id="'block-data-'+b.hash"></v-json>
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </td>
+          </tr>
+        </tbody>
+      </table>
+    </div>
+  </div>
+
+  <div v-else-if="page.block!==undefined">
+    <v-block :block="page.block" title="Block" :route="route">
+    </v-block>
+  </div>
+
+  <div v-else>
+    Page not found
+  </div>
+
+</div>
