+++ linera-explorer/html/block.html
@@ -1,0 +1,116 @@
+<div class="card">
+  <div class="card-body">
+    <h5 class="card-title">
+      <span>{{ title }}</span>
+      <button class="btn btn-link btn-sm" data-bs-toggle="modal" :data-bs-target="'#json-head-'+block.hash">
+        <i class="bi bi-braces"></i>
+      </button>
+      <div :id="'json-head-'+block.hash" class="modal fade">
+        <div class="modal-dialog modal-xl">
+          <div class="modal-content">
+            <div class="modal-header">
+              <h5 class="modal-title">Head</h5>
+              <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
+            </div>
+            <div class="modal-body">
+              <v-json :data="block" id="head-data"></v-json>
+            </div>
+          </div>
+        </div>
+      </div>
+    </h5>
+    <ul class="list-group">
+      <li class="list-group-item d-flex justify-content-between">
+        <span><strong>Hash</strong></span>
+        {{ block.hash }}
+      </li>
+      <li class="list-group-item d-flex justify-content-between">
+        <span><strong>Epoch</strong></span>
+        <span>{{ block.value.executedBlock.block.epoch }}</span>
+      </li>
+      <li class="list-group-item d-flex justify-content-between">
+        <span><strong>Height</strong></span>
+        <span>{{ block.value.executedBlock.block.height }}</span>
+      </li>
+      <li class="list-group-item d-flex justify-content-between">
+        <span><strong>Timestamp</strong></span>
+        <span>{{ (new Date(block.value.executedBlock.block.timestamp/1000)).toLocaleString() }}</span>
+      </li>
+      <li class="list-group-item d-flex justify-content-between">
+        <span><strong>Signer</strong></span>
+        <span>{{ block.value.executedBlock.block.authenticatedSigner }}</span>
+      </li>
+      <li class="list-group-item d-flex justify-content-between">
+        <span><strong>Previous Block</strong></span>
+        <a v-if="block.value.executedBlock.block.previousBlockHash!==undefined" @click="route('block', [['hash', block.value.executedBlock.block.previousBlockHash]])" class="btn btn-link">{{ block.value.executedBlock.block.previousBlockHash }}</a>
+        <span v-else>--</span>
+      </li>
+      <li class="list-group-item d-flex justify-content-between">
+        <span><strong>State Hash</strong></span>
+        <span>{{ block.value.executedBlock.stateHash }}</span>
+      </li>
+      <li class="list-group-item d-flex justify-content-between">
+        <span><strong>Status</strong></span>
+        <span>{{ block.value.status }}</span>
+      </li>
+      <li v-if="block.value.executedBlock.block.incomingMessages.length!==0" class="list-group-item d-flex justify-content-between" data-bs-toggle="collapse" :data-bs-target="'#in-messages-collapse-'+block.hash">
+        <span><strong>Incoming Messages</strong> ({{ block.value.executedBlock.block.incomingMessages.length }})</span>
+        <i class="bi bi-caret-down-fill"></i>
+      </li>
+      <li v-else class="list-group-item d-flex justify-content-between">
+        <span><strong>Incoming Messages</strong> (0)</span>
+        <span></span>
+      </li>
+      <div v-if="block.value.executedBlock.block.incomingMessages.length!==0" class="collapse" :id="'in-messages-collapse-'+block.hash">
+        <ul class="list-group">
+          <li v-for="(m, i) in block.value.executedBlock.block.incomingMessages" class="list-group-item p-0">
+            <div class="card">
+              <div class="card-header">Message {{ i+1 }}</div>
+              <div class="card-body">
+                <v-json :data="m" :id="'head-inmessage-'+i"></v-json>
+              </div>
+            </div>
+          </li>
+        </ul>
+      </div>
+      <li v-if="block.value.executedBlock.messages.length!==0" class="list-group-item d-flex justify-content-between" data-bs-toggle="collapse" :data-bs-target="'#out-messages-collapse-'+block.hash">
+        <span><strong>Outgoing Messages</strong> ({{ block.value.executedBlock.messages.length }})</span>
+        <i class="bi bi-caret-down-fill"></i>
+      </li>
+      <li v-else class="list-group-item d-flex justify-content-between">
+        <span><strong>Outgoing Messages</strong> (0)</span>
+        <span></span>
+      </li>
+      <div v-if="block.value.executedBlock.messages.length!==0" class="collapse" :id="'out-messages-collapse-'+block.hash">
+        <ul class="list-group">
+          <li v-for="(m, i) in block.value.executedBlock.messages" class="list-group-item p-0">
+            <div class="card">
+              <div class="card-header">Message {{ i+1 }}</div>
+              <div class="card-body">
+                <v-json :data="m" :id="'head-outmessage-'+i"></v-json>
+              </div>
+            </div>
+          </li>
+        </ul>
+      </div>
+      <li v-if="block.value.executedBlock.block.operations.length!==0" class="list-group-item d-flex justify-content-between" data-bs-toggle="collapse" :data-bs-target="'#operations-collapse-'+block.hash">
+        <span><strong>Operations</strong> ({{ block.value.executedBlock.block.operations.length }})</span>
+        <i class="bi bi-caret-down-fill"></i>
+      </li>
+      <li v-else class="list-group-item d-flex justify-content-between">
+        <span><strong>Operations</strong> (0)</span>
+        <span></span>
+      </li>
+      <div v-if="block.value.executedBlock.block.operations.length!==0" class="collapse" :id="'operations-collapse-'+block.hash">
+        <ul class="list-group">
+          <li v-for="(o, i) in block.value.executedBlock.block.operations" class="list-group-item p-0">
+            <div class="card card-body">
+              <div class="card-title">Operation {{ i+1 }}</div>
+              <v-op :op="o" :id="'head-'+i"></v-op>
+            </div>
+          </li>
+        </ul>
+      </div>
+    </ul>
+  </div>
+</div>
