+++ linera-explorer/js/block.js
@@ -1,0 +1,6 @@
+import template from '../html/block.html'
+
+export default {
+  template,
+  props: ["block","title", "route"]
+}
