+++ linera-explorer/js/operation.js
@@ -1,0 +1,6 @@
+import template from '../html/operation.html'
+
+export default {
+  template,
+  props: ["op","id"],
+}
