+++ linera-explorer/webpack.config.js
@@ -1,31 +1,17 @@
 const path = require("path");
 const CopyPlugin = require("copy-webpack-plugin");
 const WasmPackPlugin = require("@wasm-tool/wasm-pack-plugin");
-
 const dist = path.resolve(__dirname, "dist");
 
 module.exports = {
   mode: "production",
-  entry: {
-    index: "./js/index.js"
-  },
-  output: {
-    path: dist,
-    filename: "[name].js"
-  },
-  devServer: {
-    contentBase: dist,
-  },
+  entry: { index: "./js/index.js" },
+  output: { path: dist, filename: "[name].js" },
+  devServer: { contentBase: dist },
   plugins: [
-    new CopyPlugin({
-      patterns: [
-        { from: "static", to: "." }
-      ]}),
-    new WasmPackPlugin({
-      crateDirectory: __dirname,
-    }),
+    new CopyPlugin({ patterns: [ { from: "index.html", to: "." } ]}),
+    new WasmPackPlugin({ crateDirectory: __dirname }),
   ],
-  experiments: {
-    asyncWebAssembly: true
-  }
+  experiments: { asyncWebAssembly: true },
+  module: { rules: [ { test: /\.html$/i, loader: "html-loader" } ] },
 };
